Change: moved desktop_session to resource provider
