+++ providers/resource/bin/desktop_session_resource.py
