+++ providers/resource/jobs/resource.pxu
@@ -532,3 +532,9 @@ plugin: resource
 _summary: Discover if the system supports power modes via acpi
 _description: Discover resource info from /sys/firmware/apci
 command: platform_profile_resource.py
+
+id: desktop_session
+plugin: resource
+_summary: Check whether a desktop session is available and of which type.
+environ: XDG_SESSION_TYPE XDG_CURRENT_DESKTOP
+command: desktop_session_resource.py resources
